+++ ShooterGame/Plugins/AccelByteUe4Sdk/Source/AccelByteUe4Sdk/Private/Core/AccelByteRegistry.cpp
@@ -18,7 +18,9 @@
 #include "Api/AccelByteQosManagerApi.h"
 #include "Api/AccelByteQos.h"
 #include "Api/AccelByteLeaderboardApi.h"
+#include "Api/AccelByteCloudSaveApi.h"
 #include "Api/AccelByteGameTelemetryApi.h"
+#include "Api/AccelByteAgreementApi.h"
 #include "GameServerApi/AccelByteServerOauth2Api.h"
 #include "GameServerApi/AccelByteServerDSMApi.h"
 #include "GameServerApi/AccelByteServerStatisticApi.h"
@@ -46,9 +48,11 @@ Api::Lobby FRegistry::Lobby(FRegistry::Credentials, FRegistry::Settings);
 Api::GameProfile FRegistry::GameProfile(FRegistry::Credentials, FRegistry::Settings);
 Api::Statistic FRegistry::Statistic(FRegistry::Credentials, FRegistry::Settings);
 Api::QosManager FRegistry::QosManager(FRegistry::Credentials, FRegistry::Settings);
+Api::CloudSave FRegistry::CloudSave(FRegistry::Credentials, FRegistry::Settings);
 Api::Qos FRegistry::Qos;
 Api::Leaderboard FRegistry::Leaderboard(FRegistry::Credentials, FRegistry::Settings);
 Api::GameTelemetry FRegistry::GameTelemetry(FRegistry::Credentials, FRegistry::Settings);
+Api::Agreement FRegistry::Agreement(FRegistry::Credentials, FRegistry::Settings);
 GameServerApi::ServerOauth2 FRegistry::ServerOauth2(FRegistry::ServerCredentials, FRegistry::ServerSettings);
 GameServerApi::ServerDSM FRegistry::ServerDSM(FRegistry::ServerCredentials, FRegistry::ServerSettings);
 GameServerApi::ServerStatistic FRegistry::ServerStatistic(FRegistry::ServerCredentials, FRegistry::ServerSettings);